backend/requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aioredis==2.0.1
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.7.14
celery==5.4.0
charset-normalizer==3.4.2
Deprecated==1.2.18
distro==1.9.0
dj-database-url==2.1.0
Django==5.2.4
django-cors-headers==4.7.0
django-redis==6.0.0
djangorestframework==3.16.0
frozenlist==1.7.0
gunicorn==23.0.0
idna==3.10
magic-filter==1.0.12
multidict==6.6.3
netaddr==1.3.0
packaging==25.0
pillow==11.3.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
redis==6.2.0
requests==2.32.4
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wrapt==1.17.2
yarl==1.20.1
yookassa==3.6.0
